allocate: use unsigned long to record allocations
Smatch does a ton of ptr_list allocations and the frees aren't accounted
for if you free things one at a time, so these ints can wrap around.

+++ allocate.h
@@ -14,7 +14,7 @@ struct allocator_struct {
 	unsigned int chunking;
 	void *freelist;
 	/* statistics */
-	unsigned int allocations, total_bytes, useful_bytes;
+	unsigned long allocations, total_bytes, useful_bytes;
 };
 
 struct allocator_stats {
